Fleshing out the controllers, started a people::search

+++ models/crm_person.hpp
@@ -2,17 +2,17 @@
 #include "model.hpp"
 #include "model_factory.hpp"
 
-#define CRM_PERSON_COLUMNS                         \
-  COLUMN(firstname,   std::string, "varchar(128)") \
-  COLUMN(lastname,    std::string, "varchar(128)") \
-  COLUMN(distinction, std::string, "varchar(128)") \
-  COLUMN(sex_id,      int,				 "int")          \
-  COLUMN(language_id, int,				 "int")          \
-  COLUMN(email,       std::string, "varchar(128)") \
-  COLUMN(phone,       std::string, "varchar(128)") \
-  COLUMN(active,      int,         "int")          \
-  COLUMN(created_at,  std::tm,     "datetime")     \
-  COLUMN(updated_at,  std::tm,     "datetime")
+#define CRM_PERSON_COLUMNS                           \
+  COLUMN(firstname,   std::string,   "varchar(128)") \
+  COLUMN(lastname,    std::string,   "varchar(128)") \
+  COLUMN(distinction, std::string,   "varchar(128)") \
+  COLUMN(sex_id,      long long int, "int")          \
+  COLUMN(language_id, long long int, "int")          \
+  COLUMN(email,       std::string,   "varchar(128)") \
+  COLUMN(phone,       std::string,   "varchar(128)") \
+  COLUMN(active,      int,           "int")          \
+  COLUMN(created_at,  std::tm,       "datetime")     \
+  COLUMN(updated_at,  std::tm,       "datetime")
 
 class CrmPerson : public Model::Instance<CrmPerson> { 
   public:
